requirements.txt
pandas>=1.3.0
python-pptx>=0.6.21
openpyxl>=3.0.9
lxml>=4.6.0 